add edit click modal load with guest data available
Please note the approach taken involves a re-rendering of the entire party-planning page. Consequently, I've had to redeclare several consts for DOM elements to allow DOM manipulation in the promise chain following the rendering. This needs to be revisited as it's not particularly clean/elegant, but works for the time being.